refactor: project_context_management MCP
